scripts/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.55.3
httpx==0.27.2
tenacity==8.2.3
requests==2.31.0

